BE/fix: 계좌 이체 or 충전 메서드 금액 필드 Integer -> Long

+++ Backend/mydata/src/main/java/com/taesan/mydata/domain/bank/entity/Account.java
@@ -44,11 +44,11 @@ public class Account {
 
     private Integer lastPaidInCnt;
 
-    public void withdraw(int amt) {
+    public void withdraw(Long amt) {
         balanceAmt -= amt;
     }
 
-    public void deposit(int amt) {
+    public void deposit(Long amt) {
         balanceAmt += amt;
     }
